Init room-based connection
