+++ package.json
@@ -18,6 +18,7 @@
     "mongoose": "^5.11.12",
     "multer": "^1.4.2",
     "multer-gridfs-storage": "^4.2.0",
+    "nodemon": "^2.0.7",
     "socket.io": "^4.0.0",
     "validator": "^13.5.2"
   },
